Change type of 'balance' field in Account and Transaction to BigDecimal in order to use hibernate for precision after point

+++ src/main/java/com/account/manager/am/service/CustomerService.java
@@ -2,8 +2,10 @@ package com.account.manager.am.service;
 
 import com.account.manager.am.model.Customer;
 
+import java.math.BigDecimal;
+
 public interface CustomerService {
 
     Customer getCustomerById(int customerId);
-    void saveNewAccount(int customerId, double initialCredit);
+    void saveNewAccount(int customerId, BigDecimal initialCredit);
 }
